Bugfix: when clicking on name input, overlay closes
Clicks were bubbling up from the input box in the name overlay, and if you already had a name in the box, clicking on it would bubble up to the overlay, which would then closed. Added a barrier on the overlay so none of the child clicks would bubble up to it.